+++ tools/src/test/java/it/eng/tools/configuration/OkHttpClientConfigurationTest.java
@@ -2,7 +2,6 @@ package it.eng.tools.configuration;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
 
 import java.security.KeyManagementException;
 import java.security.KeyStoreException;
@@ -14,15 +13,12 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.MockedStatic;
-import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.boot.ssl.NoSuchSslBundleException;
 import org.springframework.boot.ssl.SslBundles;
 
+import it.eng.tools.ssl.ocsp.OcspProperties;
 import okhttp3.OkHttpClient;
 
 @ExtendWith(MockitoExtension.class)
@@ -30,6 +26,9 @@ class OkHttpClientConfigurationTest {
 
     @InjectMocks
     private OkHttpClientConfiguration configuration;
+    
+    @Mock
+    private OcspProperties ocspProperties;
 
     @Mock
     private SslBundles sslBundles;