+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] - 13-03-2025
+
+### Added
+
+ - OCSP logic for validating TLS certificate
+ - New properties for ocsp logic
+ - TLS properties for connector; default connector setup is http
+
 ## [0.1.1] - 03-03-2025
 
 ### Added